+++ CHANGES.rst
@@ -32,6 +32,9 @@ Specviz2d
 API Changes
 -----------
 
+- Extend unix style wildcard matching for searches in UI dropdown menus and
+  API objects that support multiselect. [#3744, #3752]
+
 Cubeviz
 ^^^^^^^